+++ Datamigratie.Common/Services/Det/Models/DetZaak.cs
@@ -37,6 +37,11 @@ namespace Datamigratie.Common.Services.Det.Models
         public DateTimeOffset WijzigDatumTijd { get; set; }
         public DetZaaktype? Zaaktype { get; set; }
         public List<DetDocument>? Documenten { get; set; }
+        public List<DetBesluit>? Besluiten { get; set; }
+        public string? RedenStart { get; set; }
+        public DetArchiveerGegevens? ArchiveerGegevens { get; set; }
+        public DetGeolocatie? Geolocatie { get; set; }
+        public DetKanaal? Kanaal { get; set; }
     }
 
     public class DetDocument
@@ -44,7 +49,7 @@ namespace Datamigratie.Common.Services.Det.Models
         public required List<DetDocumentVersie> DocumentVersies { get; set; }
         public string? Kenmerk { get; set; }
         public required string Titel { get; set; }
-        public string? Beschrijving {get; set; }
+        public string? Beschrijving { get; set; }
         public DocumentTaal? Taal { get; set; }
         public DocumentVorm? DocumentVorm { get; set; }
 
@@ -69,6 +74,12 @@ namespace Datamigratie.Common.Services.Det.Models
         public decimal? Bedrag { get; set; }
         public string? Betaalstatus { get; set; }
         public string? Kenmerk { get; set; }
+        public DateOnly? TransactieDatum { get; set; }
+    }
+
+    public class DetArchiveerGegevens
+    {
+        public DateOnly? BewaartermijnEinddatum { get; set; }
     }
 
     public class DocumentTaal
@@ -77,6 +88,36 @@ namespace Datamigratie.Common.Services.Det.Models
     }
 
     public class DocumentVorm
+    {
+        public required string Naam { get; set; }
+    }
+
+    public class DetBesluit
+    {
+        public required string FunctioneleIdentificatie { get; set; }
+        public required DetBesluittype Besluittype { get; set; }
+        public required DateOnly BesluitDatum { get; set; }
+        public DateOnly? Vervaldatum { get; set; }
+        public DateOnly? Ingangsdatum { get; set; }
+        public DateOnly? Reactiedatum { get; set; }
+        public DateOnly? Publicatiedatum { get; set; }
+        public string? Toelichting { get; set; }
+    }
+
+    public class DetBesluittype
+    {
+        public required string Naam { get; set; }
+        public string? Omschrijving { get; set; }
+        public bool Actief { get; set; }
+    }
+  
+    public class DetGeolocatie
+    {
+        public string? Type { get; set; }
+        public List<decimal>? Point2D { get; set; }
+    }
+
+    public class DetKanaal
     {
         public required string Naam { get; set; }
     }