+++ Datamigratie.Server/Config/ServiceCollectionExtensions.cs
@@ -23,6 +23,7 @@ using Datamigratie.Server.Features.MigrateZaken.ManageMigrations.StartMigration.
 using Datamigratie.Server.Features.Map.GlobalMapping.DocumentstatusMapping.Save.Services;
 using Datamigratie.Server.Features.MigrateZaken.ManageMigrations.StartMigration.ValidateMappings.DocumentProperty;
 using Datamigratie.Server.Features.MigrateZaken.ManageMigrations.StartMigration.ValidateMappings.Vertrouwelijkheid;
+using Datamigratie.Server.Features.MigrateZaken.ManageMigrations.StartMigration.ValidateMappings.Besluittype;
 
 namespace Datamigratie.Server.Config
 {
@@ -55,6 +56,7 @@ namespace Datamigratie.Server.Config
             services.AddScoped<IShowDocumentPropertyMappingsService, ShowDocumentPropertyMappingsService>();
             services.AddScoped<ISaveDocumentPropertyMappingsService, SaveDocumentPropertyMappingsService>();
             services.AddScoped<IValidateVertrouwelijkheidMappingsService, ValidateVertrouwelijkheidMappingsService>();
+            services.AddScoped<IValidateBesluittypeMappingsService, ValidateBesluittypeMappingsService>();
 
             services.AddHostedService<StartMigrationBackgroundService>();
             services.AddSingleton<IMigrationBackgroundTaskQueue>(ctx =>
@@ -62,7 +64,7 @@ namespace Datamigratie.Server.Config
                 return new MigrationBackgroundTaskQueue(MigrationBackgroundTaskQueueCapacity);
             });
             services.AddSingleton<MigrationWorkerState>();
-            
+
             return services;
 
         }