+++ docs/changelog/changelog.md
@@ -22,3 +22,6 @@
 - [Validate zaaknummer length per Zaaktype before migration](https://dimpact.atlassian.net/browse/DATA-92)
 - [Global mapping & migration of Documentstatussen](https://dimpact.atlassian.net/browse/DATA-195)
 - [Mapping & migration Vertrouwelijkheid per Zaaktype](https://dimpact.atlassian.net/browse/DATA-181)
+
+- [Migrate Besluit of Zaak](https://dimpact.atlassian.net/browse/DATA-192)
+- [Migrate RSIN](https://dimpact.atlassian.net/browse/DATA-162)